restructure eth
