--- modules/anonify-eth-driver/src/eth/client.rs
@@ -1,175 +1,0 @@
-use super::primitives::{Web3Contract, Web3Http};
-use crate::{error::Result, eventdb::EventCache, traits::*, utils::*, workflow::*};
-
-use frame_common::{state_types::UpdatedState, traits::*};
-use sgx_types::sgx_enclave_id_t;
-use std::{path::Path, sync::Arc};
-use web3::types::{Address, TransactionReceipt};
-use parking_lot::RwLock;
-
-/// Components needed to deploy a contract
-#[derive(Debug)]
-pub struct EthDeployer {
-    enclave_id: sgx_enclave_id_t,
-    web3_conn: Web3Http,
-    address: Option<Address>, // contract address
-}
-
-impl Deployer for EthDeployer {
-    fn new(enclave_id: sgx_enclave_id_t, node_url: &str) -> Result<Self> {
-        let web3_conn = Web3Http::new(node_url)?;
-
-        Ok(EthDeployer {
-            enclave_id,
-            web3_conn,
-            address: None,
-        })
-    }
-
-    fn get_account(&self, index: usize, password: &str) -> Result<Address> {
-        self.web3_conn.get_account(index, password)
-    }
-
-    fn deploy<P: AsRef<Path>>(
-        &mut self,
-        host_output: host_output::JoinGroup,
-        confirmations: usize,
-        abi_path: P,
-        bin_path: P,
-    ) -> Result<String> {
-        let contract_addr =
-            self.web3_conn
-                .deploy(host_output, confirmations, abi_path, bin_path)?;
-        self.address = Some(contract_addr);
-
-        Ok(hex::encode(contract_addr.as_bytes()))
-    }
-
-    // TODO: generalize, remove abi.
-    fn get_contract<P: AsRef<Path>>(self, abi_path: P) -> Result<ContractKind> {
-        let addr = self
-            .address
-            .expect("The contract hasn't be deployed yet.")
-            .to_string();
-        let contract_info = ContractInfo::new(abi_path, &addr);
-        Ok(ContractKind::Web3Contract(Web3Contract::new(
-            self.web3_conn,
-            contract_info,
-        )?))
-    }
-
-    fn get_enclave_id(&self) -> sgx_enclave_id_t {
-        self.enclave_id
-    }
-
-    fn get_node_url(&self) -> &str {
-        &self.web3_conn.get_eth_url()
-    }
-}
-
-/// Components needed to send a transaction
-#[derive(Debug)]
-pub struct EthSender {
-    enclave_id: sgx_enclave_id_t,
-    contract: Web3Contract,
-}
-
-impl Sender for EthSender {
-    fn new<P: AsRef<Path>>(
-        enclave_id: sgx_enclave_id_t,
-        node_url: &str,
-        contract_info: ContractInfo<'_, P>,
-    ) -> Result<Self> {
-        let web3_http = Web3Http::new(node_url)?;
-        let contract = Web3Contract::new(web3_http, contract_info)?;
-
-        Ok(EthSender {
-            enclave_id,
-            contract,
-        })
-    }
-
-    fn from_contract(enclave_id: sgx_enclave_id_t, contract: ContractKind) -> Self {
-        match contract {
-            ContractKind::Web3Contract(contract) => EthSender {
-                enclave_id,
-                contract,
-            },
-        }
-    }
-
-    fn get_account(&self, index: usize, password: &str) -> Result<Address> {
-        self.contract.get_account(index, password)
-    }
-
-    fn send_report_handshake(
-        &self,
-        host_output: host_output::JoinGroup,
-        confirmations: usize,
-        method: &str,
-    ) -> Result<TransactionReceipt> {
-        self.contract
-            .send_report_handshake(host_output, confirmations, method)
-    }
-
-    fn send_instruction(
-        &self,
-        host_output: host_output::Instruction,
-        confirmations: usize,
-    ) -> Result<TransactionReceipt> {
-        self.contract.send_instruction(host_output, confirmations)
-    }
-
-    fn handshake(
-        &self,
-        host_output: host_output::Handshake,
-        confirmations: usize,
-    ) -> Result<TransactionReceipt> {
-        self.contract.handshake(host_output, confirmations)
-    }
-
-    fn get_contract(self) -> ContractKind {
-        ContractKind::Web3Contract(self.contract)
-    }
-}
-
-/// Components needed to watch events
-pub struct EventWatcher {
-    contract: Web3Contract,
-    cache: Arc<RwLock<EventCache>>,
-}
-
-impl Watcher for EventWatcher {
-    fn new<P: AsRef<Path>>(
-        node_url: &str,
-        contract_info: ContractInfo<'_, P>,
-        cache: Arc<RwLock<EventCache>>,
-    ) -> Result<Self> {
-        let web3_http = Web3Http::new(node_url)?;
-        let contract = Web3Contract::new(web3_http, contract_info)?;
-
-        Ok(EventWatcher {
-            contract,
-            cache,
-        })
-    }
-
-    fn block_on_event<S: State>(
-        &self,
-        eid: sgx_enclave_id_t,
-    ) -> Result<Option<Vec<UpdatedState<S>>>> {
-        let enclave_updated_state = self
-            .contract
-            .get_event(self.cache.clone(), self.contract.address())?
-            .into_enclave_log()?
-            .verify_order()?
-            .insert_enclave(eid)?
-            .save_cache(self.contract.address());
-
-        Ok(enclave_updated_state.updated_states())
-    }
-
-    fn get_contract(self) -> ContractKind {
-        ContractKind::Web3Contract(self.contract)
-    }
-}
