+++ modules/anonify-eth-driver/src/eth/mod.rs
@@ -1,4 +1,8 @@
-pub mod client;
+pub mod deployer;
+pub mod event_watcher;
+pub mod sender;
 pub mod primitives;
 
-pub use self::client::{EthDeployer, EthSender, EventWatcher};
+pub use self::deployer::EthDeployer;
+pub use self::event_watcher::EventWatcher;
+pub use self::sender::EthSender;
