+++ modules/anonify-eth-driver/src/lib.rs
@@ -3,10 +3,10 @@
 pub mod dispatcher;
 mod error;
 pub mod eth;
-mod eventdb;
+mod cache;
 pub mod traits;
 mod utils;
 mod workflow;
 
 pub use dispatcher::Dispatcher;
-pub use eventdb::EventCache;
+pub use cache::EventCache;
